MaaS: Use maas_external_ip_address properly
The `maas_external_ip_address` variable was ignored in the `rpc_maas`
role. This patch sets the variable to `external_lb_vip_address` as a
default and allows a deployer to override the default if needed.

Manual backport of 0ffd8af from master. The master branch has some
Ansible 2 updates that aren't needed for Mitaka.

Connects rcbops/u-suk-dev#1036